Fix for disabling "Format specifies type 'int' but the argument has type 'NSUInteger' (aka 'unsigned long')" warning

+++ Classes/ObjectMappings/JSReportDescriptor.m
@@ -18,7 +18,7 @@
 @synthesize attachments = _attachments;
 
 - (NSString *)description {
-    return [NSString stringWithFormat:@"Report Descriptor - uuid: %@; Original URI: %@; Total Pages: %@; Start Page: %@; End Page: %@; Attachments Count: %i", self.uuid, self.originalUri, self.totalPages, self.startPage, self.endPage, self.attachments.count];
+    return [NSString stringWithFormat:@"Report Descriptor - uuid: %@; Original URI: %@; Total Pages: %@; Start Page: %@; End Page: %@; Attachments Count: %lu", self.uuid, self.originalUri, self.totalPages, self.startPage, self.endPage, (unsigned long)self.attachments.count];
 }
 
 @end
